+++ crates/application/Cargo.toml
@@ -11,6 +11,9 @@ publish = false
 [lints]
 workspace = true
 
+[features]
+bench-utils = []
+
 [dependencies]
 # Workspace
 constantinople-mempool.workspace = true
@@ -53,3 +56,13 @@ criterion.workspace = true
 [[bench]]
 name = "executor"
 harness = false
+
+[[bench]]
+name = "compute"
+harness = false
+required-features = ["bench-utils"]
+
+[[bench]]
+name = "lazy_body_prepare"
+harness = false
+required-features = ["bench-utils"]
