+++ crates/application/benches/compute.rs
@@ -1,0 +1,3 @@
+fn main() {
+    constantinople_application::consensus::bench::compute();
+}
