+++ crates/application/benches/lazy_body_prepare.rs
@@ -1,0 +1,3 @@
+fn main() {
+    constantinople_application::consensus::bench::lazy_body_prepare();
+}
